requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.0.6
django-bootstrap-form==3.4
django-filter==24.2
django-heroku==0.3.1
gunicorn==22.0.0
numpy==2.0.0
packaging==24.1
pillow==10.3.0
psycopg2==2.9.9
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
whitenoise==6.7.0
